pyproject.toml
[tool.poetry]
name = "diplom-11"
version = "0.1.0"
description = ""
authors = ["BlknDm <89090018300@mail.ru>"]
readme = "README.md"
packages = [{include = "diplom_11"}]

[tool.poetry.dependencies]
python = "^3.10"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"
django = "^4.1.7"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.1.0"
django-filter = "^23.1"
pydantic = "^1.10.7"
requests = "^2.28.2"
pytest = "^7.3.1"
six = "^1.16.0"
python-dateutil = "^2.8.2"
faker = "^18.5.1"
factory-boy = "^3.2.1"
inflection = "^0.5.1"
pytest-factoryboy = "^2.5.1"
pytest-django = "^4.5.2"


[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = 'todolist.settings'
python_files = [
    'tests.py',
    'test_*.py'
]
